+++ core/android/src/main/kotlin/com/moengage/flutter/MoEInitializer.kt
@@ -1,9 +1,9 @@
 package com.moengage.flutter
 
 import android.content.Context
-import com.moengage.core.Logger
 import com.moengage.core.MoEngage
-import com.moengage.core.model.IntegrationMeta
+import com.moengage.core.internal.logger.Logger
+import com.moengage.core.internal.model.IntegrationMeta
 import com.moengage.flutter.BuildConfig.MOENGAGE_FLUTTER_LIBRARY_VERSION
 import com.moengage.plugin.base.PluginInitializer.initialize
 
@@ -22,7 +22,22 @@ class MoEInitializer {
                 initialize(
                     context,
                     builder,
-                    IntegrationMeta(INTEGRATION_TYPE, MOENGAGE_FLUTTER_LIBRARY_VERSION)
+                    true
+                )
+            } catch (e: Exception) {
+                Logger.e("$tag initialize() : Exception: ", e)
+            }
+        }
+
+        @JvmStatic
+        fun initialize(context: Context, builder: MoEngage.Builder, isSdkEnabled: Boolean) {
+            try {
+                Logger.v("$tag initialize() : Will try to initialize the sdk.")
+                initialize(
+                    context,
+                    builder,
+                    IntegrationMeta(INTEGRATION_TYPE, MOENGAGE_FLUTTER_LIBRARY_VERSION),
+                    isSdkEnabled
                 )
             } catch (e: Exception) {
                 Logger.e("$tag initialize() : Exception: ", e)