+++ inbox/CHANGELOG.md
@@ -1,5 +1,10 @@
 # MoEngage Flutter Inbox Plugin
 
+### Version 2.0.0
+
+- Android Native SDK updated to support version `11.0.03` and above.
+- Plugin Base `2.0.00`
+
 ### 1.0.2 *(15th February 2021)*
 - Android artifacts use manven central instead of Jcenter.
 - Native SDK version `5.3.1`